src/requirements.txt
asgiref==3.11.0
Django==5.2.10
pillow==12.1.0
python-dotenv==1.0.1
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
psycopg2-binary==2.9.9
streamlit==1.40.0
supabase==2.10.0
